Move shared DelegatingAggregator out of OldDoFn
